Moved cards into a json file
