--- webapp/cards.py
@@ -1,637 +1,0 @@
-import random
-
-
-# does what it says on the tin
-def shuffle_deck(deck):
-    random.shuffle(deck)
-
-
-# pick card(s) - call this method multiple times to draw multiple unique cards from the deck
-def get_card(deck):
-    num = random.randint(0, len(deck) - 1)  # the minus 1 fixes the zero indexed array out-of-range error
-    card = deck[num]
-    del (deck[num])  # so we don't get the same card twice if we're calling this multiple times for the same hand
-    rev = random.randint(-1, 1)  # is card reversed? zero (false) or 1 (true)
-    drawn = (card, rev)  # tuple of card dictionary + 1 or zero for reversal
-    return drawn
-
-
-# array of dicts for each card
-def get_deck():
-    deck = [
-
-        {"name": "The Magician", "url": "the_magician", "image": "images/01.jpeg",
-         "desc": "The start of something. Beginner’s luck. Having various tools and means at our disposal. Use of supernatural forces. Creating reality with mind power. Training and acquisition of practical skills. Improvisation. Display or show for other people.",
-         "message": "create a new reality",
-         "rdesc": "Trickery, sleight of hand, cheating. Showing off, pretending. Lack of self-awareness about body, sexuality, or basic motives. Near miss due to inexperience or inaccuracy.",
-         "sequence": 1,
-         "qabalah": "from Kether to Tiphareth",
-         "hebrew_letter": "&#1488;",
-         "meditation" : "The wind blows where it wills, and you hear the sound of it\nbut you do not know whence it comes or whither it goes;\nso it is with every one who is born of the Spirit.\n(John 3:8)",
-         "cardtype": "major"},
-
-        {"name": "The Popess", "url": "the_popess", "image": "images/02.jpeg",
-         "desc": "Wisdom combining intellect and intuition. A spiritual mother. A woman hiding her strengths in a world of men. Modesty. Secrets, something hidden, mystery. Getting a hint of something which remains largely unknown. Impossible to give a definite answer now.",
-         "message": "know how to set boundaries",
-         "rdesc": "Need to hide our true nature behind the conventions of normal society. Conservative approach to sex and the body. Emotional blockage.",
-         "sequence": 2,
-         "qabalah": "from Kether to Chokmah",
-         "hebrew_letter": "&#1489;",
-         "meditation" : "Wisdom has built her house,\nshe has set up seven pillars.\n(Proverbs 9:1)",
-         "cardtype": "major"},
-
-        {"name": "The Empress", "url": "the_empress", "image": "images/03.jpeg",
-         "desc": "Abundance, growth, productivity. Natural or human touch within an artificial framework. Emotional intelligence. Protection and care. Motherhood. A powerful female figure. Strong feminine identity.",
-         "message": "act from the guts",
-         "rdesc": "Impulsive behavior, someone difficult to reason with. Over-protectiveness, excessive involvement in the life of others. Problems with a strong mother figure.",
-         "sequence": 3,
-         "qabalah": "from Kether to Binah",
-         "hebrew_letter": "&#1490;",
-         "meditation" : "Behold the handmaid of the Lord;\nlet it be unto me according to your word.\n(Luke 1:38)",
-         "cardtype": "major"},
-
-        {"name": "The Emperor", "url": "the_emperor", "image": "images/04.jpeg",
-         "desc": "Practical and material achievements. Matters relating to the workplace or source of income. Authority and control, a commanding position. A protective father figure, patron or sponsor. Assertiveness. Military affairs.",
-         "message": "show leadership and responsibility",
-         "rdesc": "Belligerence, violence, trying to solve things by brute force. Dictatorship. Possibility of sexual abuse. Difficulty in coping with a dominant father figure. Denial and hiding of inner weaknesses.",
-         "sequence": 4,
-         "qabalah": "between Chokmah and Binah",
-         "hebrew_letter": "&#1491;",
-         "meditation" : "Blessed is he who comes in the name of the Lord.\n(Luke 13:35)",
-         "cardtype": "major"},
-
-        {"name": "The Pope", "url": "the_pope", "image": "images/05.jpeg",
-         "desc": "Teacher, instructor, or counselor. Education and knowledge, academic expertise. Organized religion, conventional medicine or psychology. Spiritual father. Consultation or treatment by a specialist. Marriage.",
-         "message": "respect knowledge and education",
-         "rdesc": "Excessive adhesion to conventions and outdated norms. Bureaucracy, an oppressive establishment. Hypocrisy, discrimination. Divorce.",
-         "sequence": 5,
-         "qabalah": "from Tiphareth to Chokmah",
-         "hebrew_letter": "&#1492;",
-         "meditation" : "I am the way, and the truth, and the life.\nNo-one comes to the Father but by me.\n(John 14:6)",
-         "cardtype": "major"},
-
-        {"name": "The Lover", "url": "the_lover", "image": "images/06.jpeg",
-         "desc": "Love, amorous relationship. Emotional entanglement. Need to make a choice, or to disengage oneself from past influences. Inclinations of the heart correspond to the will of heaven. Small steps actually taken are the visible signs of inner desire.",
-         "message": "follow the path of the heart.",
-         "rdesc": "Complex relationship between several people, e.g., a romantic triangle or a tension between mother and wife. Hesitation, quandary. Confusion as to one’s own feeling and will.",
-         "sequence": 6,
-         "qabalah": "from Binah to Tiphareth",
-         "hebrew_letter": "&#1493;",
-         "meditation" : "I, Wisdom, dwell in prudence,\nand I posess knowledge and discretion...\nI love those who love me,\nand those who seek me find me.\n(Proverbs 7:12,17)",
-         "cardtype": "major"},
-
-        {"name": "The Chariot", "url": "the_chariot", "image": "images/07.jpeg",
-         "desc": "Victory or an achievement putting the querent in a strong and protected position. Ambition, energy, motivation to move forward. Public honor. Power and high status.",
-         "message": "dare and win",
-         "rdesc": "Inner weakness hidden behind external show-off. Arrogance, vanity. Over-protectiveness, emotional closure. Confusion about one’s goals. Losing the simple touch with people and reality.",
-         "sequence": 7,
-         "qabalah": "from Binah to Geburah",
-         "hebrew_letter": "&#1494;",
-         "meditation" : "Then the devil left him, and behold,\nAngels came and ministered to him.\n(Matther 3:11)",
-         "cardtype": "major"},
-
-        {"name": "Justice", "url": "justice", "image": "images/08.jpeg",
-         "desc": "Law and order, legal and court issues. A fair and balanced judgment. A developed conscience. Rationality, reasoning by clear rules and common norms. Touch of grace and humanity beyond the objective considerations.",
-         "message": "act with reason and by the accepted norms",
-         "rdesc": "Petty accountability, a critical and judgmental attitude, guilt feelings. Repressive control of self and of others. Negative ideas blocking change and advance.",
-         "sequence": 8,
-         "qabalah": "from Chesed to Chokmah",
-         "hebrew_letter": "&#1495;",
-         "meditation" : "Behold my servant, whom I have chosen; my beloved,\nin whom my soul is well pleased:\nI will put my spirit upon him,\nand he shall show justice to the gentiles.\n(Matthew 12:18)",
-         "cardtype": "major"},
-
-        {"name": "The Hermit", "url": "the_hermit", "image": "images/09.jpeg",
-         "rdesc": "A closed and reclusive attitude. Isolation, loneliness. Fixed ideas. Excessive caution and suspicion, a critical approach looking for defects. Hidden and denied desires.",
-         "message": "look for the essence of things",
-         "desc": "A quest for truth or for spiritual understanding. Concentrating on a clear purpose. Caution, careful examination. Self-privation for the sake of a meaningful cause. Loyalty to principles, strong faith.",
-         "sequence": 9,
-         "qabalah": "between Geburah and Chesed",
-         "hebrew_letter": "&#1496;",
-         "meditation" : "For the gate is narrow and the way is hard that leads to life\nand those who find it are few.\n(Matthew 7:14)",
-         "cardtype": "major"},
-
-        {"name": "The Wheel of Fortune", "url": "the_wheel_of_fortune", "image": "images/10.jpeg",
-         "desc": "Change in circumstances and position. A rise after a fall. Gambling, putting faith in capricious luck. Life cycles, closure of circles. Adapting to the routine of everyday life. A hint to previous incarnations.",
-         "message": "accept life’s ups and downs",
-         "rdesc": "A decline after a period of rising. Danger lurks at the summit. Moving in a closed circle. Capricious mood changes. Feeling powerless to affect one’s situation.",
-         "sequence": 10,
-         "qabalah": "from Tiphareth to Chesed",
-         "hebrew_letter": "&#1497;",
-         "meditation" : "Vanity of vanities, all is vanity...\nWhat has been is what will be,\nand what has been done is what will be done;\nand there is nothing new under the sun.\n(Ecclesiastes 1:2,9)",
-         "cardtype": "major"},
-
-        {"name": "Strength", "url": "strength", "image": "images/11.jpeg",
-         "desc": "Power and courage to face challenges. Controlled expression of creative urges, drives and desires. Mobilization of inner resources towards a common goal. Taking risks.",
-         "message": "take control of yourself",
-         "rdesc": "The need to keep things under control leads to constant tensions. A risk of losing one’s grip. Internal conflicts and unrealistic assessment of one’s own forces may lead to failure.",
-         "sequence": 11,
-         "qabalah": "from Geburah to Tiphareth",
-         "hebrew_letter": "&#1499;",
-         "meditation" : "This is the strongest of all powers, the force of all forces;\nit overcomes every subtle thing and penetrates every solid.\n(the Emerald Tablet, 9)",
-         "cardtype": "major"},
-
-        {"name": "The Hanged Man", "url": "the_hanged_man", "image": "images/12.jpeg",
-         "desc": "Seeing things from a unique point of view. Enduring difficulties for a worthy cause. A period of deep self examination. Passivity, acceptance of reality even if it is the opposite of what one expects.",
-         "message": "look at things from the opposite perspective",
-         "rdesc": "Isolation. Emotional stance of a victim. Inability to act. Denying one’s own unique qualities, striving to be “normal” at all costs. Living in one’s private and imaginary reality.",
-         "sequence": 12,
-         "qabalah": "from Netzach to Chesed",
-         "hebrew_letter": "&#1500;",
-         "meditation" : "Foxes have holes, and birds of the air have nests;\nbut the Son of man has nowhere to lay his head.\n(Matthew 8:20)",
-         "cardtype": "major"},
-
-        {"name": "13", "url": "13", "image": "images/13.jpeg",
-         "desc": "The end of something whose time has come. Cutting off past influences or attachment to dominant figures. Giving up the superfluous and keeping only the essential. Disintegration of the old makes room for the new.",
-         "message": "give up what is over",
-         "rdesc": "Difficulty in coping with loss or change. Temporary difficulties, a trying challenge. Disintegration. Realisation of a painful truth. Does not predict future death, but may reflect anxiety about dying or mourning over a loss which has already happened.",
-         "sequence": 13,
-         "qabalah": "from Hod to Geburah",
-         "hebrew_letter": "&#1502;",
-         "meditation" : "And the woman said to the serpent \"We may eat of the\nfruit of the trees of the garden, but God said 'You shall\nnot eat of the fruit of the tree which is in the midst of the\ngarden, neither shall you touch it, lest you die.'\"\nBut the serpent said to the woman \"You will not die.\nFor God knows that when you eat of it your eyes will be\nopened, and you will be as gods, knowing good and evil.\"\n(Genesis 3:2-5)",
-         "cardtype": "major"},
-
-        {"name": "Temperance", "url": "temperance", "image": "images/14.jpeg",
-         "desc": "Reconciliation, compromise, relaxation of tensions. Integration of opposites. Ability to do the seemingly impossible. A slow process of distillation and improvement. Patience, perseverance. Self-improvement.",
-         "message": "find the golden mean",
-         "rdesc": "Going back and forth without making real progress. Losing patience with a lengthy process. Emotional preoccupation with oneself, pushing away others who might come to help.",
-         "sequence": 14,
-         "qabalah": "from Netzach to Tiphareth",
-         "hebrew_letter": "&#1504;",
-         "meditation" : "Everyone who drinks of this water will thirst again,\nbut whoever drinks of the water that I shall give him\nwill never thirst; the water that I shall give him will\nbecome in him a spring of water welling up to eternal life.\n(John 4:13-14)",
-         "cardtype": "major"},
-
-        {"name": "The Devil", "url": "the_devil", "image": "images/15.jpeg",
-         "desc": "A burst of creativity. Paradoxes and contradictions. Irony and mocking of common norms. Acting from desires, passions and impulses. Moving on from a past family trauma.",
-         "message": "express passion and desire.",
-         "rdesc": "Temptation, attraction to the dark and forbidden. Exploitation, egotism, domination. Compulsive self-gratification. Senseless behavior has its price. Difficulty in detaching oneself from an unhealthy bond.",
-         "sequence": 15,
-         "qabalah": "from Tiphareth to Hod",
-         "hebrew_letter": "&#1505;",
-         "meditation" : "",
-         "cardtype": "major"},
-
-        {"name": "The House of God", "url": "the_house_of_god", "image": "images/16.jpeg",
-         "desc": "Breaking up of solid structures. Getting free from confinement. Sudden breakthrough after long preparations. Sparkling sexual encounter. Success lies in simplicity and modesty.",
-         "message": "return to the solid ground of reality",
-         "rdesc": "Shock, collapse of projects or trusted structures. A fall from an apparently solid and secure position. Chaos, confusion, difficulty in understanding what is going on. Vanity and pride lead to failure.",
-         "sequence": 16,
-         "qabalah": "between Netzach and Hod",
-         "hebrew_letter": "&#1506;",
-         "meditation" : "He who exalts himself will be abased,\nand he who humbles himself will be exalted.\n(Luke 14:11)",
-         "cardtype": "major"},
-
-        {"name": "The Star", "url": "the_star", "image": "images/17.jpeg",
-         "desc": "Openness, simplicity, return to nature. Purity, honesty. Showing yourself “as you are,” accepting one’s body and desires. Generosity. Luck from heaven. Intuitive feeling of guidance or energy coming from a higher plane.",
-         "message": "flow from a pure source",
-         "rdesc": "Naive optimism and wishful-thinking. Exposing oneself to danger or abuse. Difficulty in setting proper boundaries. Squandering, wastefulness.",
-         "sequence": 17,
-         "qabalah": "between Hod and Yesod",
-         "hebrew_letter": "&#1508;",
-         "meditation" : "The righteous flourish like the palm tree,\nand grow like a cedar in Lebanon...\nThey still bring forth fruit in old age,\nthey are ever full of sap and green...\n(Psalm 92:12-14)",
-         "cardtype": "major"},
-
-        {"name": "The Moon", "url": "the_moon", "image": "images/18.jpeg",
-         "desc": "Deep emotions, perhaps related to a mother or feminine figure. A different experience of reality. Longing for the unreachable. Finding one’s hidden strengths. Occupation with the remote past. A hidden treasure.",
-         "message": "don’t be afraid to go deep down",
-         "rdesc": "Vague and disturbing feelings. Emotional difficulties, a period of depression. Danger lurking under the surface. Retreat, the road ahead is hard to find.",
-         "sequence": 18,
-         "qabalah": "between Yesod and Netzach",
-         "hebrew_letter": "&#1510;",
-         "meditation" : "The intellect is characterised by a naturai inability to\ncomprehend life... But it is to the very inwardness of\nlife that intuition leads us — by intuition I mean instinct\nthat has become disinterested, self-conscious, capable\nof reflecting upon its object and of enlarging it indefinitely.\n(Henri Bergson, Creative Evolution)",
-         "cardtype": "major"},
-
-        {"name": "The Sun", "url": "the_sun", "image": "images/19.jpeg",
-         "desc": "Light and warmth, abundance, blessings. Pleasant feeling, emotional or physical healing. Partnership, trust, sharing, brotherhood. Human touch. An ideal father figure. Matters relating to children. Setting limits in a moderate and non-oppressive way.",
-         "message": "find suitable partners",
-         "rdesc": "Living in a limited space, difficulty to face reality “in the open.” Immaturity, dependence on others. Someone or something too intense and energetic to feel comfortable with. An absent father.",
-         "sequence": 19,
-         "qabalah": "between Tiphareth and Yesod",
-         "hebrew_letter": "&#1511;",
-         "meditation" : "I am the Alpha and the Omega,\nthe first and the last,\nthe beginning and the end.\n(Revelation 22:13)",
-         "cardtype": "major"},
-
-        {"name": "Judgement", "url": "judgement", "image": "images/20.jpeg",
-         "desc": "Revelation, enlightenment, a new understanding. A turning point in a therapy process. Healing of a family relationship. Disclosure, secrets revealed, publicity. Birth of a baby or of a new thing.",
-         "message": "awaken to spiritual reality",
-         "rdesc": "Revelation of something that should have been kept hidden. Lack of privacy. Unpleasant realization. Problems related to child-parent relations. Too much noise and drama.",
-         "sequence": 20,
-         "qabalah": "between Yesod and Malkuth",
-         "hebrew_letter": "&#1512;",
-         "meditation" : "For as the Father raises the dead and gives them life,\nso also the Son gives life to whom he will.\nThe Father judges no one,\nbut has given all judgement to the Son...\n(John 5:21-22)",
-         "cardtype": "major"},
-
-        {"name": "The Fool", "url": "the_fool", "image": "images/00.jpeg",
-         "desc": "Freedom from conventions and norms. Something or someone unique and exceptional. Options kept open. Giving up control, spontaneity. Uncertainty, attention to the here and now. Going on a trip.",
-         "message": "keep on the move",
-         "rdesc": "Difficulty in choosing and committing oneself to something stable. Restlessness. Lack of purpose. Getting lost. Foolish behavior. Eccentricity, lack of acceptance by the social environment. Difficulty in planning ahead.",
-         "sequence": 21,
-         "qabalah": "between Malkuth and Netzach",
-         "hebrew_letter": "&#64299;",
-         "meditation" : "If any one among you thinks that he is wise in this age,\nlet him become a fool that he may become wise.\nFor the wisdom of this world is folly before God.\n(I Corinthians 3:18-19)",
-         "cardtype": "major"},
-
-        {"name": "The World", "url": "the_world", "image": "images/21.jpeg",
-         "desc": "Completion of a process. Balanced activity and achievements in various domains. Contact with far places. Harmony and correspondence between different planes. Pregnancy, something new is about to be born. The Dance of Life.",
-         "message": "everything is perfect as it is",
-         "rdesc": "Life in a bubble, difficulty in sharing your world with others. Disconnection of inner feelings from external life. Preoccupation with oneself, idealized self-image, inability to move forward.",
-         "sequence": 22,
-         "qabalah": "between Hod and Malkuth",
-         "hebrew_letter": "&#1514;",
-         "meditation" : "Then I was at work beside him,\nand I was daily his delight,\nrejoicing before him always,\nrejoicing in his inhabited world\nand delighting in the sons of men.\n(Proverbs 7 30-31)",
-         "cardtype": "major"},
-
-        {"name": "King of Swords", "url": "king_of_swords", "image": "images/swki.jpeg",
-         "desc": "A determination to break from the past, a strong will. Feeling equipped to deal with uncertainty. Wisdom and intellectual maturity.",
-         "rdesc": "A divided heart, a need to cut off from something to which one is still attached. Over-calculating in a vain attempt to overcome uncertainty.",
-         "sequence": 23,
-         "qabalah": "&#1497; of &#x1f701; (Yod of Air)",
-         "cardtype": "court"},
-
-        {"name": "Queen of Swords", "url": "queen_of_swords", "image": "images/swqu.jpeg",
-         "desc": "A secure and protected position. Defending one’s territory. Preparation of something that shouldn’t be exposed yet.",
-         "rdesc": "Defensiveness and rigidity. Suspicion and fixed ideas block advancement and preclude new connections.",
-         "sequence": 24,
-         "qabalah": "&#1492; of &#x1f701; (Heh of Air)",
-         "cardtype": "court"},
-
-        {"name": "Knight of Swords", "url": "knight_of_swords", "image": "images/swkn.jpeg",
-         "desc": "Energy and resources to advance, still looking for the right direction. Hovering above practical constraints. Determination and perseverance.",
-         "rdesc": "Trying to force one’s misguided views, insisting on a wrong direction. Losing touch with the ground.",
-         "sequence": 25,
-         "qabalah": "&#1493; of &#x1f701; (Vav of Air)",
-         "cardtype": "court"},
-
-        {"name": "Page of Swords", "url": "page_of_swords", "image": "images/swpa.jpeg",
-         "desc": "Preparation for a future challenge. Looking for a compromise between reason and strong desires. Hesitating to use one’s power.",
-         "rdesc": "Confusion, negative and inhibiting thoughts, self-defeat. Sloppy use of one’s own tools may cause damage.",
-         "sequence": 26,
-         "qabalah": "&#1492; of &#x1f701; (Heh (final) of Air)",
-         "cardtype": "court"},
-
-        {"name": "Ace of Swords", "url": "ace_of_swords", "image": "images/sw01.jpeg",
-         "desc": "A planned initiative. Rational and logical thinking, sharpness of mind. A conclusive decision. Readiness to fight. Ambition, competitiveness. A victory with stable achievements.",
-         "rdesc": "Negative and unproductive thoughts. Misconceptions, delusions. Self-defeat. Injury.",
-         "sequence": 27,
-         "qabalah": "&#x1f701; (Air) of Kether",
-         "cardtype": "ace"},
-
-        {"name": "Two of Swords", "url": "two_of_swords", "image": "images/sw02.jpeg",
-         "desc": "Boundaries. Limits that protect and define something which is in development. Making full use of the present situation. Preparations for future advancement. A clear view encompassing the overall situation.",
-         "rdesc": "(similar to upright) Boundaries. Limits that protect and define something which is in development. Making full use of the present situation. Preparations for future advancement. A clear view encompassing the overall situation.",
-         "sequence": 28,
-         "qabalah": "☾ in ♎︎ (Moon in Libra)",
-         "cardtype": "minor"},
-
-        {"name": "Three of Swords", "url": "three_of_swords", "image": "images/sw03.jpeg",
-         "desc": "Victory. Overcoming a weak opposition. Cutting through a quandary and going forward in a clear direction. A third party intervenes and wins over two weakened opponents.",
-         "rdesc": "A failure, defeat from a weaker opponent. An unsuccessful attempt to make a decisive move.",
-         "sequence": 29,
-         "qabalah": "♄ in ♎︎ (Saturn in Libra)",
-         "cardtype": "minor"},
-
-        {"name": "Four of Swords", "url": "four_of_swords", "image": "images/sw04.jpeg",
-         "desc": "Restriction. A limited space for development and maneuver. Trying to push against constraints. Potentials to grow once the present limitations become less solid.",
-         "rdesc": "Confinement and blocking, lack of motivation or energy to break out of a limited situation.",
-         "sequence": 30,
-         "qabalah": "♃ in ♎︎ (Jupiter in Libra)",
-         "cardtype": "minor"},
-
-        {"name": "Five of Swords", "url": "five_of_swords", "image": "images/sw05.jpeg",
-         "desc": "Breakthrough. A forward thrust overcoming the existing limits. Keeping up spirits in a tight situation. Doing things one’s own way.",
-         "rdesc": "A failed initiative to change the situation. Stubbornness leading nowhere. Oppressing factors cannot be removed now.",
-         "sequence": 31,
-         "qabalah": "♀ in ♒︎ (Venus in Aquarius)",
-         "cardtype": "minor"},
-
-        {"name": "Six of Swords", "url": "six_of_swords", "image": "images/sw06.jpeg",
-         "desc": "Adaptation. Accepting limitations and adapting oneself to them. Respecting the present order. Compromising in order to make the best of the existing situation.",
-         "rdesc": "Resignation, surrender, giving up the ambition to change things for the better. Lack of fighting spirit.",
-         "sequence": 32,
-         "qabalah": "☿ in ♒︎ (Mercury in Aquarius)",
-         "cardtype": "minor"},
-
-        {"name": "Seven of Swords", "url": "seven_of_swords", "image": "images/sw07.jpeg",
-         "desc": "Sharpness. A focused and determined attitude. Concentrating on a clear goal and doing what it takes to reach it. Winning a fight with the odds evenly balanced.",
-         "rdesc": "A narrow and over-concentrated vision. Investing one’s efforts and resources in a lost cause.",
-         "sequence": 33,
-         "qabalah": "☾ in ♒︎ (Moon in Aquarius)",
-         "cardtype": "minor"},
-
-        {"name": "Eight of Swords", "url": "eight_of_swords", "image": "images/sw08.jpeg",
-         "desc": "Defenses. Putting up shields and blocks. Psychological defense mechanisms. A need to be in total control. A well-guarded treasure. Entering another’s domain with permission.",
-         "rdesc": "(similar to upright) Defenses. Putting up shields and blocks. Psychological defense mechanisms. A need to be in total control. A well-guarded treasure. Entering another’s domain with permission.",
-         "sequence": 34,
-         "qabalah": "♃ in ♊︎ (Jupiter in Gemini)",
-         "cardtype": "minor"},
-
-        {"name": "Nine of Swords", "url": "nine_of_swords", "image": "images/sw09.jpeg",
-         "desc": "Courage. Winning a fight against a superior force. Pure intentions. Putting imperfect means to good use.",
-         "rdesc": "Losing against a stronger opponent. Sloppiness, imperfect preparations for a challenge.",
-         "sequence": 35,
-         "qabalah": "♂ in ♊︎ (Mars in Gemini)",
-         "cardtype": "minor"},
-
-        {"name": "Ten of Swords", "url": "ten_of_swords", "image": "images/sw10.jpeg",
-         "desc": "Exhaustion. A complex situation with many conflicting interests. A long battle without a clear outcome. Need to find an ally who will attack the problem from a different direction.",
-         "rdesc": "Immobility. Impossible to move now. Feeling attacked from different sides. A painful and humiliating defeat.",
-         "sequence": 36,
-         "qabalah": "☉ in ♊︎ (Sun in Gemini)",
-         "cardtype": "minor"},
-
-        {"name": "King of Coins", "url": "king_of_coins", "image": "images/peki.jpeg",
-         "desc": "Confidence and security, a cautious but optimistic vision. Looking for new achievements while holding existing assets secure.",
-         "rdesc": "Dissatisfaction with what one already has. Disregard of the good things in the present situation. A limited outlook.",
-         "sequence": 37,
-         "qabalah": "&#1497; of &#x1f703; (Yod of Earth)",
-         "cardtype": "court"},
-
-        {"name": "Queen of Coins", "url": "queen_of_coins", "image": "images/pequ.jpeg",
-         "desc": "Tangible assets, material and personal stability, a sober and realistic vision. Looking at things from a practical and pragmatic perspective.",
-         "rdesc": "Conservatism, resistance to change, aiming only to preserve the existing assets. Looking at things only from the material perspective.",
-         "sequence": 38,
-         "qabalah": "&#1492; of &#x1f703; (Heh of Earth)",
-         "cardtype": "court"},
-
-        {"name": "Knight of Coins", "url": "knight_of_coins", "image": "images/pekn.jpeg",
-         "desc": "Advancement in a practical direction. A productive expression of creativity. A clear goal in sight.",
-         "rdesc": "Constant pursuit of money, without reaching material stability. Passions and desires may interfere with practical plans.",
-         "sequence": 39,
-         "qabalah": "&#1493; of &#x1f703; (Vav of Earth)",
-         "cardtype": "court"},
-
-        {"name": "Page of Coins", "url": "page_of_coins", "image": "images/pepa.jpeg",
-         "desc": "A practical endeavor. Untapped potentials are within reach. Tangible success at the beginning. A solid material base for further advancement.",
-         "rdesc": "Hesitation, lack of clear purpose. Thinking in terms of past achievements misses present opportunities.",
-         "sequence": 40,
-         "qabalah": "&#1492; of &#x1f703; (Heh (final) of Earth)",
-         "cardtype": "court"},
-
-        {"name": "Ace of Coins", "url": "ace_of_coins", "image": "images/pe01.jpeg",
-         "desc": "A good start in material things. Financial and physical stability. A practical perspective. A significant sum of money. Utilitarian approach. Greed. Something basic and unsophisticated.",
-         "rdesc": "(similar to upright) A good start in material things. Financial and physical stability. A practical perspective. A significant sum of money. Utilitarian approach. Greed. Something basic and unsophisticated.",
-         "sequence": 41,
-         "qabalah": "&#x1f703; (Earth) of Kether",
-         "cardtype": "ace"},
-
-        {"name": "Two of Coins", "url": "two_of_coins", "image": "images/pe02.jpeg",
-         "desc": "Duality. Two options or two elements. Collaborating while keeping distance. A winding road, advancing in complex ways. Recognition and acknowledgment.",
-         "rdesc": "(similar to upright) Duality. Two options or two elements. Collaborating while keeping distance. A winding road, advancing in complex ways. Recognition and acknowledgment.",
-         "sequence": 42,
-         "qabalah": "♃ in ♑︎ (Jupiter in Capricorn)",
-         "cardtype": "minor"},
-
-        {"name": "Three of Coins", "url": "three_of_coins", "image": "images/pe03.jpeg",
-         "desc": "Product. A partnership or an alliance bears fruit. First results of a project. Good prospects.",
-         "rdesc": "Disappointment, a partnership or a project does not bear the expected fruits.",
-         "sequence": 43,
-         "qabalah": "♂ in ♑︎ (Mars in Capricorn)",
-         "cardtype": "minor"},
-
-        {"name": "Four of Coins", "url": "four_of_coins", "image": "images/pe04.jpeg",
-         "desc": "Stability. Solid material assets. Tradition, reputation and honor. Time-tested reliability. Established social institutions.",
-         "rdesc": "Conservatism, clinging to old and outdated patterns.",
-         "sequence": 44,
-         "qabalah": "☉ in ♑︎ (Sun in Capricorn)",
-         "cardtype": "minor"},
-
-        {"name": "Five of Coins", "url": "five_of_coins", "image": "images/pe05.jpeg",
-         "desc": "Disruption. Something new appears and destabilizes existing structures. A new element gets attention, but also awakens resistance.",
-         "rdesc": "(similar to upright) Disruption. Something new appears and destabilizes existing structures. A new element gets attention, but also awakens resistance.",
-         "sequence": 45,
-         "qabalah": "☿ in ♉︎ (Mercury in Taurus)",
-         "cardtype": "minor"},
-
-        {"name": "Six of Coins", "url": "six_of_coins", "image": "images/pe06.jpeg",
-         "desc": "Expansion. Abundance of resources and possible ways to advance. A positive outlook, success. A good balance between stability and movement.",
-         "rdesc": "(similar to upright) Expansion. Abundance of resources and possible ways to advance. A positive outlook, success. A good balance between stability and movement.",
-         "sequence": 46,
-         "qabalah": "☾ in ♉︎ (Moon in Taurus)",
-         "cardtype": "minor"},
-
-        {"name": "Seven of Coins", "url": "seven_of_coins", "image": "images/pe07.jpeg",
-         "desc": "Acceptance. Something new is well received. Help and protection. Integrating into a system without losing one’s individuality.",
-         "rdesc": "Lack of independence, need to rely on help and acceptance from others.",
-         "sequence": 47,
-         "qabalah": "♄ in ♉︎ (Saturn in Taurus)",
-         "cardtype": "minor"},
-
-        {"name": "Eight of Coins", "url": "eight_of_coins", "image": "images/pe08.jpeg",
-         "desc": "Uniformity. A mechanical structure. Practical considerations prove efficient, but lack a human touch. Routine work. A slow and patient advance.",
-         "rdesc": "(similar to upright) Uniformity. A mechanical structure. Practical considerations prove efficient, but lack a human touch. Routine work. A slow and patient advance.",
-         "sequence": 48,
-         "qabalah": "☉ in ♍︎ (Sun in Virgo)",
-         "cardtype": "minor"},
-
-        {"name": "Nine of Coins", "url": "nine_of_coins", "image": "images/pe09.jpeg",
-         "desc": "Motivation. Carving a niche for oneself in an existing system. Ambition. Endurance and independent thinking bear long-term fruits.",
-         "rdesc": "(similar to upright) Motivation. Carving a niche for oneself in an existing system. Ambition. Endurance and independent thinking bear long-term fruits.",
-         "sequence": 49,
-         "qabalah": "♀ in ♍︎ (Venus in Virgo)",
-         "cardtype": "minor"},
-
-        {"name": "Ten of Coins", "url": "ten_of_coins", "image": "images/pe10.jpeg",
-         "desc": "Abundance. Intensive activity in practical affairs. Material success and achievements. Some may be getting more than others.",
-         "rdesc": "(similar to upright) Abundance. Intensive activity in practical affairs. Material success and achievements. Some may be getting more than others.",
-         "sequence": 50,
-         "qabalah": "☿ in ♍︎ (Mercury in Virgo)",
-         "cardtype": "minor"},
-
-        {"name": "King of Cups", "url": "king_of_cups", "image": "images/cuki.jpeg",
-         "desc": "Emotional maturity, optimism, ability to overcome past injuries and look ahead. Openness to new things, but with prudence and caution. Closing one’s ear to voices from the past.",
-         "rdesc": "Difficulty in overcoming an emotional blow. A pessimistic outlook caused by negative past experiences. ",
-         "sequence": 51,
-         "qabalah": "&#1497; of &#x1f704; (Yod of Water)",
-         "cardtype": "court"},
-
-        {"name": "Queen of Cups", "url": "queen_of_cups", "image": "images/cuqu.jpeg",
-         "desc": "A rich inner world which is kept hidden. Guarding one’s privacy or valuable assets. Strong feelings held under control.",
-         "rdesc": "Closure, defensiveness. Distrust of others due to negative past experiences. Hiding one’s emotions under guise of rational criticism.",
-         "sequence": 52,
-         "qabalah": "&#1492; of &#x1f704; (Heh of Water)",
-         "cardtype": "court"},
-
-        {"name": "Knight of Cups", "url": "knight_of_cups", "image": "images/cukn.jpeg",
-         "desc": "A romantic gesture, offering one’s heart, courting. Openness, sincerity, a simple heart. A potential lover may appear.",
-         "rdesc": "Superficial and unstable feelings. An over-optimistic but unrealistic attitude. An overt display of shallow or insincere feelings. ",
-         "sequence": 53,
-         "qabalah": "&#1493; of &#x1f704; (Vav of Water)",
-         "cardtype": "court"},
-
-        {"name": "Page of Cups", "url": "page_of_cups", "image": "images/cupa.jpeg",
-         "desc": "First and unsure steps in a romantic endeavor. Shyness. Sincere intentions. Trying to figure out one’s feelings.",
-         "rdesc": "Over-absorption in one’s personal feelings, losing contact with others. Sloppiness in practical affairs.",
-         "sequence": 54,
-         "qabalah": "&#1492; of &#x1f704; (Heh (final) of Water)",
-         "cardtype": "court"},
-
-        {"name": "Ace of Cups", "url": "ace_of_cups", "image": "images/cu01.jpeg",
-         "desc": "The beginning of a love relationship. Expression of warm feelings. Romantic longing for something extraordinary. Emotional and spiritual growth.",
-         "rdesc": "Emotional dryness, feeling oneself empty. Avoidance of intimacy, negative feelings, heartbreak.",
-         "sequence": 55,
-         "qabalah": "&#x1f704; (Water) of Kether",
-         "cardtype": "ace"},
-
-        {"name": "Two of Cups", "url": "two_of_cups", "image": "images/cu02.jpeg",
-         "desc": "Partnership. A romantic relationship or a close personal alliance. Interpersonal dynamics based on social norms. Passion in a love relationship, which may turn against itself.",
-         "rdesc": "A crisis in a couple relationship, disappointment with someone close to you.",
-         "sequence": 56,
-         "qabalah": "♀ in ♋︎ (Venus in Cancer)",
-         "cardtype": "minor"},
-
-        {"name": "Three of Cups", "url": "three_of_cups", "image": "images/cu03.jpeg",
-         "desc": "Birth. Something new brings joy and happiness. Caring for a child. Issues of child-parent relations. A common project motivated by feelings and not only by interests.",
-         "rdesc": "Problems in relations with one’s parents, or with one’s child. A strong alliance of two persons leaves a third one outside.",
-         "sequence": 57,
-         "qabalah": "☿ in ♋︎ (Mercury in Cancer)",
-         "cardtype": "minor"},
-
-        {"name": "Four of Cups", "url": "four_of_cups", "image": "images/cu04.jpeg",
-         "desc": "Family. A collective of people (family, community etc.) with a history and a sense of belonging. Commitment to a group, at the price of giving up personal interests.",
-         "rdesc": "Problems and discord in the family or in a long-lasting community. A fixed social structure which doesn’t allow for adaptation or flexibility.",
-         "sequence": 58,
-         "qabalah": "☾ in ♋︎ (Moon in Cancer)",
-         "cardtype": "minor"},
-
-        {"name": "Five of Cups", "url": "five_of_cups", "image": "images/cu05.jpeg",
-         "desc": "Links. Popularity, relations with many people. Becoming the center of attention in a group. Relying on connections with other people to advance oneself or to overcome difficulties.",
-         "rdesc": "Excessive preoccupation with social activity. Losing oneself in multiple superficial connections. Cultivating virtual instead of real contacts.",
-         "sequence": 59,
-         "qabalah": "♂ in ♏︎ (Mars in Scorpio)",
-         "cardtype": "minor"},
-
-        {"name": "Six of Cups", "url": "six_of_cups", "image": "images/cu06.jpeg",
-         "desc": "Continuity. A long-term relationship. Repetition between different generations in the family. A stable personal alliance.",
-         "rdesc": "Monotony, tedious repetition. Falling time and again into the same emotional traps.",
-         "sequence": 60,
-         "qabalah": "☉ in ♏︎ (Sun in Scorpio)",
-         "cardtype": "minor"},
-
-        {"name": "Seven of Cups", "url": "seven_of_cups", "image": "images/cu07.jpeg",
-         "desc": "Individuality. A single person finding his place in a group. Contact with people in high positions. Exceptional qualities are appreciated.",
-         "rdesc": "Problems of integration in a group or an organization. Being part of a collective, but feeling isolated and estranged.",
-         "sequence": 61,
-         "qabalah": "♀ in ♏︎ (Venus in Scorpio)",
-         "cardtype": "minor"},
-
-        {"name": "Eight of Cups", "url": "eight_of_cups", "image": "images/cu08.jpeg",
-         "desc": "Involvement. Developing personal relationships within a group. A favorable human-relations environment. A feast or a family event.",
-         "rdesc": "Interference of the environment in a couple’s relationships. Pressures from one’s family in romantic or personal matters.",
-         "sequence": 62,
-         "qabalah": "♄ in ♓︎ (Saturn in Pisces)",
-         "cardtype": "minor"},
-
-        {"name": "Nine of Cups", "url": "nine_of_cups", "image": "images/cu09.jpeg",
-         "desc": "Collectivity. A group or organization working harmoniously with everyone finding his proper place. Accepting one’s role in a social environment. Happiness, wishes coming true.",
-         "rdesc": "A confusing social situation, difficulty in situating oneself within a complex environment.",
-         "sequence": 63,
-         "qabalah": "♃ in ♓︎ (Jupiter in Pisces)",
-         "cardtype": "minor"},
-
-        {"name": "Ten of Cups", "url": "ten_of_cups", "image": "images/cu10.jpeg",
-         "desc": "Leadership. A person with special qualities receives appreciation and high status. Assuming responsibility for others. Maintaining a superior position.",
-         "rdesc": "A fallen leader, loss of popularity. Disappointment because of ingratitude by people one has helped.",
-         "sequence": 64,
-         "qabalah": "♂ in ♓︎ (Mars in Pisces)",
-         "cardtype": "minor"},
-
-        {"name": "King of Clubs", "url": "king_of_clubs", "image": "images/waki.jpeg",
-         "desc": "A mature attitude to urges and desires. Controlled creativity. Prodding oneself to make a move forward. Investing present assets in future projects.",
-         "rdesc": "Plans to move forward are frustrated by self-defeating acts. Hesitation, conflicts, tendency to make things too heavy and complex.",
-         "sequence": 65,
-         "qabalah": "&#1497; of &#x1f702; (Yod of Fire)",
-         "cardtype": "court"},
-
-        {"name": "Queen of Clubs", "url": "queen_of_clubs", "image": "images/waqu.jpeg",
-         "desc": "A feminine figure with a strong personality. Things connected with food and eating. Speaking softly while holding a big stick. A secure, well-defended position.",
-         "rdesc": "Intimidation, menace. Using sexuality as a means of control. Problems with a strong mother figure. Fear of feminine power.",
-         "sequence": 66,
-         "qabalah": "&#1492; of &#x1f702; (Heh of Fire)",
-         "cardtype": "court"},
-
-        {"name": "Knight of Clubs", "url": "knight_of_clubs", "image": "images/wakn.jpeg",
-         "desc": "A change of direction, following one’s urges and passions. A temporary stop, but there is still energy and a desire to advance.",
-         "rdesc": "Preoccupation with the satisfaction of one’s own desires. Problem in defining long-term goals. Submitting to temptation.",
-         "sequence": 67,
-         "qabalah": "&#1493; of &#x1f702; (Vav of Fire)",
-         "cardtype": "court"},
-
-        {"name": "Page of Clubs", "url": "page_of_clubs", "image": "images/wapa.jpeg",
-         "desc": "A creative potential which still needs processing. Keeping a safe distance from events and waiting for the right moment.",
-         "rdesc": "A task too heavy for the querent’s strength. Difficulty in controlling desires and urges. Immature approach to sexuality.",
-         "sequence": 68,
-         "qabalah": "&#1492; of &#x1f702; (Heh (final) of Fire)",
-         "cardtype": "court"},
-
-        {"name": "Ace of Clubs", "url": "ace_of_clubs", "image": "images/wa01.jpeg",
-         "desc": "Creative momentum. Active sexuality. Strong impulses, energy and drive. Life force. Beginning of growth. Dispersing one’s efforts in different directions.",
-         "rdesc": "Lack of energy, restriction, repressed sexuality, a creative block.",
-         "sequence": 69,
-         "qabalah": "&#x1f702; (Fire) of Kether",
-         "cardtype": "ace"},
-
-        {"name": "Two of Clubs", "url": "two_of_clubs", "image": "images/wa02.jpeg",
-         "desc": "Crossroads. Several options or ways to choose from. Every course offers benefits. A brief encounter with someone going his own way. Blocking an opponent’s line of advance.",
-         "rdesc": "(similar to upright) Crossroads. Several options or ways to choose from. Every course offers benefits. A brief encounter with someone going his own way. Blocking an opponent’s line of advance.",
-         "sequence": 70,
-         "qabalah": "♂  in ♈︎ (Mars in Aries)",
-         "cardtype": "minor"},
-
-        {"name": "Three of Clubs", "url": "three_of_clubs", "image": "images/wa03.jpeg",
-         "desc": "Direction. Moving forward after a moment of hesitation. Finding a middle path between two courses of action. Gaining an advantage by keeping neutrality between two conflicting sides.",
-         "rdesc": "(similar to upright) Direction. Moving forward after a moment of hesitation. Finding a middle path between two courses of action. Gaining an advantage by keeping neutrality between two conflicting sides.",
-         "sequence": 71,
-         "qabalah": "☉ in ♈︎ (Sun in Aries)",
-         "cardtype": "minor"},
-
-        {"name": "Four of Clubs", "url": "four_of_clubs", "image": "images/wa04.jpeg",
-         "desc": "Stalemate. A temporary stop in order to prepare for future advancement. Tensions at present, but good prospects in the long run. Making a move now is in nobody’s interest.",
-         "rdesc": "(similar to upright) Stalemate. A temporary stop in order to prepare for future advancement. Tensions at present, but good prospects in the long run. Making a move now is in nobody’s interest.",
-         "sequence": 72,
-         "qabalah": "♀ in ♈︎ (Venus in Aries)",
-         "cardtype": "minor"},
-
-        {"name": "Five of Clubs", "url": "five_of_clubs", "image": "images/wa05.jpeg",
-         "desc": "Overcoming. Getting over a weak opposition. Breakdown of an equilibrium. Focusing on the main objective. An initiative to make a winning move.",
-         "rdesc": "(top right leaf becomes bottom left) Walking into a complex situation, losing one’s edge.",
-         "sequence": 73,
-         "qabalah": "♄ in ♌︎ (Saturn in Leo)",
-         "cardtype": "minor"},
-
-        {"name": "Six of Clubs", "url": "six_of_clubs", "image": "images/wa06.jpeg",
-         "desc": "Collaboration. A strong alliance between two parties with different goals but common present interests. A taste for luxury made possible by favorable conditions.",
-         "rdesc": "(with the decorated flower at the bottom) Excessive pursuit of luxury. Need to break up an alliance of opponents.",
-         "sequence": 74,
-         "qabalah": "♃ in ♌︎ (Jupiter in Leo)",
-         "cardtype": "minor"},
-
-        {"name": "Seven of Clubs", "url": "seven_of_clubs", "image": "images/wa07.jpeg",
-         "desc": "Struggle. Someone putting up a fight against many opponents. Obstinacy, endurance, keeping one’s position in a conflict situation. A difficult combat with an uncertain outcome.",
-         "rdesc": "(similar to upright) Struggle. Someone putting up a fight against many opponents. Obstinacy, endurance, keeping one’s position in a conflict situation. A difficult combat with an uncertain outcome.",
-         "sequence": 75,
-         "qabalah": "♂ in ♌︎ (Mars in Leo)",
-         "cardtype": "minor"},
-
-        {"name": "Eight of Clubs", "url": "eight_of_clubs", "image": "images/wa08.jpeg",
-         "desc": "Regulation. It is possible to advance only by following the rules. Occupation with short-term goals while losing the long-term perspective. A roadblock.",
-         "rdesc": "(similar to upright) Regulation. It is possible to advance only by following the rules. Occupation with short-term goals while losing the long-term perspective. A roadblock.",
-         "sequence": 76,
-         "qabalah": "☿ in ♐︎ (Mercury in Sagittarius)",
-         "cardtype": "minor"},
-
-        {"name": "Nine of Clubs", "url": "nine_of_clubs", "image": "images/wa09.jpeg",
-         "desc": "Interruption. Difficulties and oppositions too hard to overcome. Giving up one’s projects to avoid conflict. Starting anew after a challenging period.",
-         "rdesc": "(similar to upright) Interruption. Difficulties and oppositions too hard to overcome. Giving up one’s projects to avoid conflict. Starting anew after a challenging period.",
-         "sequence": 77,
-         "qabalah": "☾ in ♐︎ (Moon in Sagittarius)",
-         "cardtype": "minor"},
-
-        {"name": "Ten of Clubs", "url": "ten_of_clubs", "image": "images/wa10.jpeg",
-         "desc": "Loyalty. A partnership or an alliance endures hardships and succeeds in getting over them. Pure intentions and perseverance lead to success. Honoring one’s principles in spite of difficulties.",
-         "rdesc": "(similar to upright) Loyalty. A partnership or an alliance endures hardships and succeeds in getting over them. Pure intentions and perseverance lead to success. Honoring one’s principles in spite of difficulties.",
-         "sequence": 78,
-         "qabalah": "♄ in ♐︎ (Saturn in Sagittarius)",
-         "cardtype": "minor"}
-
-    ]
-
-    return deck
